+++ mimetic098/contentid.cxx
@@ -18,7 +18,7 @@ ContentId::ContentId()
     std::string host = gethostname();
     if(!host.length())
         host = "unknown";
-  m_cid = "c" + utils::int2str(time(0)) + "." + utils::int2str(getpid()) +
+  m_cid = "c" + utils::int2str((int)time(0)) + "." + utils::int2str(getpid()) +
         "." + utils::int2str(++ms_sequence_number) + "@" + host;
 }
