+++ mimetic098/rfc822/datetime.cxx
@@ -388,7 +388,7 @@ FieldValue* DateTime::clone() const
 
 ostream& operator<<(ostream& os, const DateTime& dt)
 {
-    int width = os.width(), fill = os.fill();
+    size_t width = os.width(), fill = os.fill();
 
     os << dt.dayOfWeek().name() << ", "
        << setw(2) << setfill('0') << dt.day() << " "
